requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
asgiref==3.7.2
astroid==2.15.5
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
cssselect==1.2.0
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
fastapi==0.97.0
filelock==3.12.2
freezegun==1.2.2
frozenlist==1.3.3
googleapis-common-protos==1.56.2
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.0.1
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
opentelemetry-api==1.18.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-exporter-jaeger-proto-grpc==1.18.0
opentelemetry-exporter-jaeger-thrift==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
packaging==23.1
platformdirs==3.6.0
pluggy==1.1.0
prometheus-client==0.17.0
protobuf==3.20.3
psycopg2-binary==2.9.6
py==1.11.0
pydantic==1.10.9
pylint==2.17.4
pyparsing==3.1.0
pyproject-api==1.5.2
pytest==7.3.2
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-stubs==0.4
starlette==0.27.0
thrift==0.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tox==4.6.2
types-ujson==5.8.0.0
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
